api/requirements.txt
# API and OpenAI integration
flask>=2.3.0
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Development tools
black>=23.0.0
pylint>=3.0.0 